move CollectResourcesAndMatch to a common lib
