+++ internal/inventory/awsfetcher/fetcher_elb_test.go
@@ -28,6 +28,7 @@ import (
 
 	"github.com/elastic/cloudbeat/internal/dataprovider/providers/cloud"
 	"github.com/elastic/cloudbeat/internal/inventory"
+	"github.com/elastic/cloudbeat/internal/inventory/testutil"
 	"github.com/elastic/cloudbeat/internal/resources/providers/awslib"
 	"github.com/elastic/cloudbeat/internal/resources/providers/awslib/elb"
 	elbv2 "github.com/elastic/cloudbeat/internal/resources/providers/awslib/elb_v2"
@@ -93,7 +94,7 @@ func TestELBv1Fetcher_Fetch(t *testing.T) {
 	identity := &cloud.Identity{Account: "123", AccountAlias: "alias"}
 	fetcher := newElbFetcher(logger, identity, providerv1, providerv2)
 
-	collectResourcesAndMatch(t, fetcher, expected)
+	testutil.CollectResourcesAndMatch(t, fetcher, expected)
 }
 
 func TestELBv2Fetcher_Fetch(t *testing.T) {
@@ -143,5 +144,5 @@ func TestELBv2Fetcher_Fetch(t *testing.T) {
 	identity := &cloud.Identity{Account: "123", AccountAlias: "alias"}
 	fetcher := newElbFetcher(logger, identity, providerv1, providerv2)
 
-	collectResourcesAndMatch(t, fetcher, expected)
+	testutil.CollectResourcesAndMatch(t, fetcher, expected)
 }
