+++ internal/inventory/awsfetcher/fetcher_iam_user_test.go
@@ -27,6 +27,7 @@ import (
 
 	"github.com/elastic/cloudbeat/internal/dataprovider/providers/cloud"
 	"github.com/elastic/cloudbeat/internal/inventory"
+	"github.com/elastic/cloudbeat/internal/inventory/testutil"
 	"github.com/elastic/cloudbeat/internal/resources/providers/awslib"
 	"github.com/elastic/cloudbeat/internal/resources/providers/awslib/iam"
 	"github.com/elastic/cloudbeat/internal/resources/utils/pointers"
@@ -128,5 +129,5 @@ func TestIAMUserFetcher_Fetch(t *testing.T) {
 	identity := &cloud.Identity{Account: "123", AccountAlias: "alias"}
 	fetcher := newIamUserFetcher(logger, identity, provider)
 
-	collectResourcesAndMatch(t, fetcher, expected)
+	testutil.CollectResourcesAndMatch(t, fetcher, expected)
 }
