+++ internal/inventory/awsfetcher/fetcher_lambda_test.go
@@ -26,6 +26,7 @@ import (
 
 	"github.com/elastic/cloudbeat/internal/dataprovider/providers/cloud"
 	"github.com/elastic/cloudbeat/internal/inventory"
+	"github.com/elastic/cloudbeat/internal/inventory/testutil"
 	"github.com/elastic/cloudbeat/internal/resources/providers/awslib"
 	"github.com/elastic/cloudbeat/internal/resources/providers/awslib/lambda"
 	"github.com/elastic/cloudbeat/internal/resources/utils/pointers"
@@ -80,5 +81,5 @@ func TestLambdaFunction_Fetch(t *testing.T) {
 	identity := &cloud.Identity{Account: "123", AccountAlias: "alias"}
 	fetcher := newLambdaFetcher(logger, identity, provider)
 
-	collectResourcesAndMatch(t, fetcher, expected)
+	testutil.CollectResourcesAndMatch(t, fetcher, expected)
 }
