+++ internal/inventory/testutil/testutil.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package awsfetcher
+package testutil
 
 import (
 	"context"
@@ -27,7 +27,7 @@ import (
 	"github.com/elastic/cloudbeat/internal/inventory"
 )
 
-func collectResourcesAndMatch(t *testing.T, fetcher inventory.AssetFetcher, expected []inventory.AssetEvent) {
+func CollectResourcesAndMatch(t *testing.T, fetcher inventory.AssetFetcher, expected []inventory.AssetEvent) {
 	t.Helper()
 
 	ch := make(chan inventory.AssetEvent)
